validation for only create connection when method existe

+++ src/backend/middleware/databaseConnection.ts
@@ -1,12 +1,15 @@
 import { getDatabaseConnector } from '@backend/database/connection';
-import { Handler, Request, Response } from '@backend/model/http';
+import { Handler, Request, RequestMethods, Response } from '@backend/model/http';
 
 const connector = getDatabaseConnector();
 
-export default () => {
+export default (methods: RequestMethods) => {
   return (fn: Handler) => async (req: Request, res: Response): Promise<void> => {
-    req.db = connector();
+    if (methods.includes(req.method)) {
+      req.db = connector();
+      await fn(req, res);
+      await req.db.destroy();
+    }
     await fn(req, res);
-    await req.db.destroy();
   };
 };
